+++ example/lib/screens/PrivateProfile.dart
@@ -38,12 +38,12 @@ import 'package:flutter/material.dart';
 
 import 'package:solid_auth/solid_auth.dart';
 
-import 'package:solid_auth_example/components/Header.dart';
+import 'package:solidautheg/components/Header.dart';
 // Project imports:
-import 'package:solid_auth_example/models/Constants.dart';
-import 'package:solid_auth_example/models/GetRdfData.dart';
-import 'package:solid_auth_example/models/SolidApi.dart' as rest_api;
-import 'package:solid_auth_example/screens/ProfileInfo.dart';
+import 'package:solidautheg/models/Constants.dart';
+import 'package:solidautheg/models/GetRdfData.dart';
+import 'package:solidautheg/models/SolidApi.dart' as rest_api;
+import 'package:solidautheg/screens/ProfileInfo.dart';
 
 class PrivateProfile extends StatefulWidget {
   final SolidAuthManager authManager;