+++ example/lib/screens/ProfileInfo.dart
@@ -39,8 +39,8 @@ import 'package:flutter/material.dart';
 import 'package:solid_auth/solid_auth.dart';
 
 // Project imports:
-import 'package:solid_auth_example/models/Constants.dart';
-import 'package:solid_auth_example/screens/EditProfile.dart';
+import 'package:solidautheg/models/Constants.dart';
+import 'package:solidautheg/screens/EditProfile.dart';
 
 class ProfileInfo extends StatelessWidget {
   final Map profData; // Profile data