+++ example/lib/screens/PublicScreen.dart
@@ -37,8 +37,8 @@ library;
 import 'package:flutter/material.dart';
 
 // Project imports:
-import 'package:solid_auth_example/models/Responsive.dart';
-import 'package:solid_auth_example/screens/PublicProfile.dart';
+import 'package:solidautheg/models/Responsive.dart';
+import 'package:solidautheg/screens/PublicProfile.dart';
 
 // ignore: must_be_immutable
 class PublicScreen extends StatelessWidget {